core/requirements.txt
Django>=3.2.4,<3.3a
psycopg2-binary>=2.8.5,<2.9a
redis==3.5.3
celery==5.2.3
pydantic==1.9.2
pillow==5.0.0
django-celery-beat==2.3.0
djangorestframework>=3.13.1,<3.14a
django-cors-headers>=3.7.0,<3.8a
requests==2.28.1
